Added new repo structure with fxml and css files
