+++ src/main/java/org/openjfx/MainApp.java
@@ -1,0 +1,47 @@
+package org.openjfx;
+
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+
+public class MainApp extends Application {
+
+    @Override
+    public void start(Stage stage) throws Exception {
+        Parent root = FXMLLoader.load(getClass().getResource("scene.fxml"));
+        
+        Scene scene = new Scene(root);
+        scene.getStylesheets().add(getClass().getResource("styles.css").toExternalForm());
+        
+        stage.setTitle("JavaFX and Gradle");
+        stage.setScene(scene);
+        stage.show();
+    }
+
+    public static void main(String[] args) {
+        launch(args);
+
+        // Boiler plate for Tests, delete later
+        for (int i = 1; i <= 100; i++) {
+            System.out.println(convert(i));
+        }
+    }
+
+    // Boiler plate for Tests, delete later
+    public static String convert(int decide) {
+        if (decide % 15 == 0) {
+            return "tutorial.HelloWorld";
+        }
+        if (decide % 3 == 0) {
+            return "Hello";
+        }
+        if (decide % 5 == 0) {
+            return "World";
+        }
+        return String.valueOf(decide);
+    }
+
+}
