+++ src/main/resources/org/openjfx/styles.css
@@ -1,0 +1,3 @@
+.label {
+    -fx-text-fill: blue;
+}
